Refactor fetchData to its own file
